Bugfix/jaicore search/mct sbugs (#232)
* Resolved bugs in MCTS and improved logging behavior of MCTSPathSearch

* initial

* Modified MCTS algorithms

* Added sandbox test

* Refined the code for PL-MCTS

* List methods

Add a function (getMethodsList) that allows to use list of required interfaces.

* Added small components test file and a small prettify method output method.

* Added ColorMaps and enabled GraphViewPlugin to show color mapped nodes

* Resolved generics issue

* fixed compile error

* updated .json component file and the prettify method method.

* Added support for "lazy" MDPs to MCTS and added this for SameGame

* Update HASCOReduction.java

El output de los métodos en caso de tener una lista de interfaces requeridas cambió de ser un String a ser una lista de VariableParam, lo propio con methodParams. Se agregó la precondición !AnyOmitted.

* Update HASCOReduction.java

Cambiar methodParams de string a list de VariableParams

* Small changes not worth writing about

* Added missing op., refactored the creation of new methods.
Testable reduction, new list-interface methods need to be fixed.

* small changes

* Update HASCOReduction.java

Changes related with the variables' names.

* All methods and ops. init task-network set to list interfaces. Not able to find solution.

* Added SameGamePath and several corrections to TAG algorithm

* Improved logging behavior in planning and added alerts in ComponentLoad

* Added graph event forwarding for HTN planners

* Updated softconf problem.

* Modified progress query

* fixed pathh

* Added changes to resolve the recent issues we had in the HASCO extension

* Updated HASCO documentation

* Resolved compile issues resulting from the modification to required list
interfaces

* Slight adjustments in the reconstruction plans of a pipeline

* made solution object explicit

* Introduced interface layer for components

* Added test compile dependency to hasco-anova

* Resolved several code smells

+++ JAICore/jaicore-search/src/main/java/ai/libs/jaicore/search/algorithms/mdp/mcts/uct/UCT.java
@@ -12,7 +12,7 @@ import ai.libs.jaicore.search.probleminputs.IMDP;
 
 public class UCT<N, A> extends MCTS<N, A> {
 
-	public UCT(final IMDP<N, A, Double> input, final double maxIterations, final double gamma, final double epsilon, final Random r, final boolean tabooExhaustedNodes) {
+	public UCT(final IMDP<N, A, Double> input, final int maxIterations, final double gamma, final double epsilon, final Random r, final boolean tabooExhaustedNodes) {
 		super(input, new UCBPolicy<>(gamma, input.isMaximizing()), new UniformRandomPolicy<>(r), maxIterations, gamma, epsilon, tabooExhaustedNodes);
 	}
